requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymssql==2.2.10
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
bcrypt==4.0.1
passlib==1.7.4
python-jose[cryptography]==3.3.0
cryptography==41.0.7
email-validator==2.1.0
